backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.32.0,<0.39
sqlalchemy>=2.0
aiosqlite>=0.19
httpx>=0.26
pydantic>=2.9
pydantic-settings>=2.1
python-dotenv>=1.0
